finish

+++ Bank.cs
@@ -5,7 +5,7 @@ namespace PlanYourHeist
 {
    public class Bank 
    {
-       public int CashOnHand {get;set;}
+       public double CashOnHand {get;set;}
        public int AlarmScore {get;set;}
        public int VaultScore {get;set;}
        public int SecurityGuardScore {get;set;}
